+++ cowichan_serial/norm.cpp
@@ -1,12 +1,12 @@
 #include "cowichan_serial.hpp"
 
-void findMinMax(PointVector points, int n, Point* minPoint, Point* maxPoint);
+void findMinMax(PointVector points, INT64 n, Point* minPoint, Point* maxPoint);
 
 void CowichanSerial::norm (PointVector pointsIn, PointVector pointsOut)
 {
   Point minPoint, maxPoint;
   real sclX, sclY; // scaling factors
-  int i;
+  INT64 i;
 
   // compute scaling factors
   findMinMax(pointsIn, n, &minPoint, &maxPoint);
@@ -31,14 +31,15 @@ void CowichanSerial::norm (PointVector pointsIn, PointVector pointsOut)
  * @param minPoint min x/y values.
  * @param maxPoint max x/y values.
  */
-void findMinMax(PointVector points, int n, Point* minPoint, Point* maxPoint) {
+void findMinMax(PointVector points, INT64 n, Point* minPoint,
+    Point* maxPoint) {
 
   minPoint->x = points[0].x;
   minPoint->y = points[0].y;
   maxPoint->x = points[0].x;
   maxPoint->y = points[0].y;
 
-  for (int i = 1; i < n; i++) {
+  for (INT64 i = 1; i < n; i++) {
     minPoint->x = std::min (points[i].x, minPoint->x);
     minPoint->y = std::min (points[i].y, minPoint->y);
     maxPoint->x = std::max (points[i].x, maxPoint->x);
